Remove duplicate entries from Schuessler's list and add variant characters
For characters that are clearly writing variants, I have considered that they
had the same pronunciation and added them to Schuessler. Perhaps another
approach would have been to build a list of character variants somewhere else;
this is still an option in the future.

This doesn't resolve the question of what to do with transcriptions of the
bronzes that are found using simplified characters in Mou & Zhong 2016,
however.